Move internal/data to model so cannot be accessed outside of model
