+++ cmd/pdatagen/internal/trace_structs.go
@@ -17,14 +17,14 @@ package internal
 var traceFile = &File{
 	Name: "trace",
 	imports: []string{
-		`otlptrace "go.opentelemetry.io/collector/internal/data/protogen/trace/v1"`,
+		`otlptrace "go.opentelemetry.io/collector/model/internal/data/protogen/trace/v1"`,
 	},
 	testImports: []string{
 		`"testing"`,
 		``,
 		`"github.com/stretchr/testify/assert"`,
 		``,
-		`otlptrace "go.opentelemetry.io/collector/internal/data/protogen/trace/v1"`,
+		`otlptrace "go.opentelemetry.io/collector/model/internal/data/protogen/trace/v1"`,
 	},
 	structs: []baseStruct{
 		resourceSpansSlice,
