+++ model/internal/data/spanid.go
