+++ model/pdata/generated_trace.go
@@ -18,7 +18,7 @@
 package pdata
 
 import (
-	otlptrace "go.opentelemetry.io/collector/internal/data/protogen/trace/v1"
+	otlptrace "go.opentelemetry.io/collector/model/internal/data/protogen/trace/v1"
 )
 
 // ResourceSpansSlice logically represents a slice of ResourceSpans.
