+++ model/pdata/logs_test.go
@@ -19,9 +19,9 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	otlpcollectorlog "go.opentelemetry.io/collector/internal/data/protogen/collector/logs/v1"
-	otlplogs "go.opentelemetry.io/collector/internal/data/protogen/logs/v1"
 	"go.opentelemetry.io/collector/model/internal"
+	otlpcollectorlog "go.opentelemetry.io/collector/model/internal/data/protogen/collector/logs/v1"
+	otlplogs "go.opentelemetry.io/collector/model/internal/data/protogen/logs/v1"
 )
 
 func TestLogRecordCount(t *testing.T) {
